pyproject.toml
[tool.poetry]
name = "dbt-interview"
version = "0.1.0"
description = "An series of exercises for an interview on dbt"
authors = ["adoherty <adoherty@kyruushealth.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
dbt-core = "^1.8.8"
dbt-duckdb = "^1.9.0"
pytest = "^8.3.4"

[tool.pytest.ini_options]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
